requirements.txt
psycopg2-binary
SQLAlchemy
pandas
streamlit
python-dotenv

